Renamed conversions to mask.
