--- cocotools/src/converters.rs
@@ -1,3 +1,0 @@
-//! Module containing utilities to convert data formats.
-
-pub mod mask;
