+++ cocotools/src/main.rs
@@ -5,8 +5,8 @@ use clap::Parser;
 
 mod argparse;
 mod coco;
-mod converters;
 mod errors;
+mod mask;
 mod utils;
 mod visualize;
 use crate::argparse::{Cli, Commands};
@@ -39,7 +39,7 @@ fn main() -> Result<(), Box<dyn error::Error>> {
             output_path,
         } => {
             let mut dataset = COCO::new(annotations_path, &PathBuf::from("N/A"))?;
-            converters::mask::convert_coco_segmentation(&mut dataset, *target_segmentation)?;
+            mask::conversions::convert_coco_segmentation(&mut dataset, *target_segmentation)?;
             let output_path = output_path
                 .as_ref()
                 .map_or_else(|| annotations_path, |output_path| output_path);
